requirements.txt
streamlit
openai
python-dotenv
Pillow
textblob
streamlit-js-eval
supabase

